Cast char into int to eliminate warning

+++ src/Params/EnvelopeParams.C
@@ -51,7 +51,7 @@ EnvelopeParams::~EnvelopeParams(){
 };
 
 REALTYPE EnvelopeParams::getdt(char i){
-    REALTYPE result=(pow(2.0,Penvdt[i]/127.0*12.0)-1.0)*10.0;//miliseconds
+    REALTYPE result=(pow(2.0,Penvdt[(int)i]/127.0*12.0)-1.0)*10.0;//miliseconds
     return(result);
 };
